refactor: extract settings logic to useSettings hook
This moves the settings logic from the SettingsPage component
into a custom hook called useSettings. This improves code
organization and reusability.
